# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hospital Command Center API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type PatientPriority = typeof PatientPriority[keyof typeof PatientPriority];


export const PatientPriority = {
  normal: 'normal',
  urgent: 'urgent',
  emergency: 'emergency',
} as const;

export type PatientStatus = typeof PatientStatus[keyof typeof PatientStatus];


export const PatientStatus = {
  waiting: 'waiting',
  called: 'called',
  consulting: 'consulting',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Patient {
  id: number;
  token: string;
  name: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  department: string;
  priority: PatientPriority;
  status: PatientStatus;
  /** @nullable */
  doctorId?: number | null;
  /** @nullable */
  doctorName?: string | null;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  estimatedWaitMinutes?: number | null;
  /** @nullable */
  queuePosition?: number | null;
  createdAt: string;
  /** @nullable */
  calledAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
}

export type PatientInputPriority = typeof PatientInputPriority[keyof typeof PatientInputPriority];


export const PatientInputPriority = {
  normal: 'normal',
  urgent: 'urgent',
  emergency: 'emergency',
} as const;

export interface PatientInput {
  name: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  department: string;
  priority: PatientInputPriority;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  doctorId?: number | null;
}

export type PatientUpdatePriority = typeof PatientUpdatePriority[keyof typeof PatientUpdatePriority];


export const PatientUpdatePriority = {
  normal: 'normal',
  urgent: 'urgent',
  emergency: 'emergency',
} as const;

export type PatientUpdateStatus = typeof PatientUpdateStatus[keyof typeof PatientUpdateStatus];


export const PatientUpdateStatus = {
  waiting: 'waiting',
  called: 'called',
  consulting: 'consulting',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface PatientUpdate {
  name?: string;
  priority?: PatientUpdatePriority;
  status?: PatientUpdateStatus;
  /** @nullable */
  doctorId?: number | null;
  department?: string;
}

export type DoctorStatus = typeof DoctorStatus[keyof typeof DoctorStatus];


export const DoctorStatus = {
  available: 'available',
  busy: 'busy',
  break: 'break',
  offline: 'offline',
} as const;

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  department: string;
  status: DoctorStatus;
  /** @nullable */
  currentPatientId?: number | null;
  /** @nullable */
  currentPatientName?: string | null;
  patientsSeenToday: number;
  avgConsultMinutes: number;
}

export interface DoctorInput {
  name: string;
  specialty: string;
  department: string;
}

export type DoctorUpdateStatus = typeof DoctorUpdateStatus[keyof typeof DoctorUpdateStatus];


export const DoctorUpdateStatus = {
  available: 'available',
  busy: 'busy',
  break: 'break',
  offline: 'offline',
} as const;

export interface DoctorUpdate {
  status?: DoctorUpdateStatus;
}

export type DepartmentLoadLevel = typeof DepartmentLoadLevel[keyof typeof DepartmentLoadLevel];


export const DepartmentLoadLevel = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
  critical: 'critical',
} as const;

export interface Department {
  id: number;
  name: string;
  code: string;
  waitingCount: number;
  activeCount: number;
  doctorsAvailable: number;
  avgWaitMinutes: number;
  loadLevel: DepartmentLoadLevel;
}

export interface QueueNextResult {
  patient?: Patient;
  message: string;
  queueEmpty: boolean;
}

export type ActivityEventType = typeof ActivityEventType[keyof typeof ActivityEventType];


export const ActivityEventType = {
  patient_added: 'patient_added',
  patient_called: 'patient_called',
  consultation_started: 'consultation_started',
  consultation_completed: 'consultation_completed',
  emergency_arrived: 'emergency_arrived',
  doctor_available: 'doctor_available',
  doctor_busy: 'doctor_busy',
} as const;

export interface ActivityEvent {
  id: number;
  type: ActivityEventType;
  message: string;
  /** @nullable */
  patientName?: string | null;
  /** @nullable */
  patientToken?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  priority?: string | null;
  timestamp: string;
}

export type InsightType = typeof InsightType[keyof typeof InsightType];


export const InsightType = {
  warning: 'warning',
  info: 'info',
  critical: 'critical',
  success: 'success',
} as const;

export interface Insight {
  id: number;
  type: InsightType;
  title: string;
  message: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  confidence?: number | null;
  createdAt: string;
}

export interface DashboardStats {
  totalWaiting: number;
  totalActive: number;
  totalCompleted: number;
  doctorsAvailable: number;
  doctorsTotal: number;
  queueVelocity: number;
  avgWaitMinutes: number;
  aiHealthScore: number;
  emergencyCount: number;
  urgentCount: number;
  /** @nullable */
  peakHourPrediction?: string | null;
  systemEfficiency: number;
}

export interface WaitingRoomToken {
  token: string;
  name: string;
  department: string;
  /** @nullable */
  doctorName?: string | null;
  status: string;
  /** @nullable */
  estimatedMinutes?: number | null;
}

export interface WaitingRoomData {
  nowServing: WaitingRoomToken[];
  nextTokens: WaitingRoomToken[];
  currentTime: string;
  totalInQueue: number;
  /** @nullable */
  estimatedClearTime?: string | null;
}