# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hospital Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEventType } from './activityEventType';

export interface ActivityEvent {
  id: number;
  type: ActivityEventType;
  message: string;
  /** @nullable */
  patientName?: string | null;
  /** @nullable */
  patientToken?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  priority?: string | null;
  timestamp: string;
}