# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hospital Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { DoctorStatus } from './doctorStatus';

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  department: string;
  status: DoctorStatus;
  /** @nullable */
  currentPatientId?: number | null;
  /** @nullable */
  currentPatientName?: string | null;
  patientsSeenToday: number;
  avgConsultMinutes: number;
}