# lib/api-zod/src/generated/types/doctorInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hospital Command Center API
 * OpenAPI spec version: 0.1.0
 */

export interface DoctorInput {
  name: string;
  specialty: string;
  department: string;
}