# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hospital Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { PatientPriority } from './patientPriority';
import type { PatientStatus } from './patientStatus';

export interface Patient {
  id: number;
  token: string;
  name: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  department: string;
  priority: PatientPriority;
  status: PatientStatus;
  /** @nullable */
  doctorId?: number | null;
  /** @nullable */
  doctorName?: string | null;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  estimatedWaitMinutes?: number | null;
  /** @nullable */
  queuePosition?: number | null;
  createdAt: string;
  /** @nullable */
  calledAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
}